Easy (almost) unique ids for ramses particles based on nbodies
The ids of the particles are offset by a value of the form:
    TYPE * 10^(ceil(log10(nbodies)+1))

In practice, for 1000 particles, the max ID is 1000 (starting at 1),
if all particles have the same type. The "uniquified" ID will be:
- gas: 1, 2, 3, etc
- DM: 1001, 1002, 1003, etc
- star: 2001, 2002, 2003, etc
- BH: 3001, 3002, 3003, etc.

Note that this only ensure that there are no two particles with the
same ID, but this dos not ensure that ID are kept constant over
multiple snapshots. For instance, if the nbodies goes to the next
power of 10 between two snapshots, all the offsets will be changed.

Note also that gas particles will not have the same IDs between two
snapshots, as ramses does not really have gas particles.

FIXME: so far, for gas particles and with MPI, the IDs are only unique
within one MPI process.